+++ avengers/presentation/view.py
@@ -1,12 +1,12 @@
 from sanic.request import Request
-from sanic.views import HTTPMethodView
 from sanic.response import HTTPResponse
+from sanic.views import HTTPMethodView
 
 from avengers.data.repositories.refresh_token import RefreshTokenRepository
 from avengers.data.repositories.unauthorized_user import UnauthorizedUserRepository
+from avengers.data.repositories.user import UserRepository
 from avengers.presentation.exceptions import InvalidSignupInfo, UserNotFound, TokenError
 from avengers.services.auth import AuthService
-from avengers.data.repositories.user import UserRepository
 
 
 class SignUp(HTTPMethodView):
@@ -17,7 +17,7 @@ class SignUp(HTTPMethodView):
 
     async def post(self, request: Request) -> HTTPResponse:
         if not request.json:
-            raise InvalidSignupInfo("")
+            raise InvalidSignupInfo()
 
         response = await self.service.sign_up(
             email=request.json.get("email"),
@@ -49,14 +49,14 @@ class UserAuth(HTTPMethodView):
 
     async def post(self, request: Request) -> HTTPResponse:
         if not request.json or not request.json.get("email") or not request.json.get("password"):
-            raise UserNotFound("")
+            raise UserNotFound()
 
         response = await self.service.login(request)
         return response
 
     async def patch(self, request: Request) -> HTTPResponse:
         if not request.headers.get("X-Refresh-Token"):
-            raise TokenError("")
+            raise TokenError()
 
         response = await self.service.refresh_token(request)
 
@@ -64,7 +64,7 @@ class UserAuth(HTTPMethodView):
 
     async def delete(self, request: Request) -> HTTPResponse:
         if not request.headers.get("X-Refresh-Token"):
-            raise TokenError("")
+            raise TokenError()
 
         response = await self.service.logout(request)
