+++ avengers/services/auth.py
@@ -4,13 +4,13 @@ from sanic_jwt_extended import create_access_token, create_refresh_token
 from werkzeug.security import check_password_hash, generate_password_hash
 
 from avengers.config import settings
-from avengers.data.repositories.refresh_token import RefreshTokenRepository
-from avengers.services import send_mail
 from avengers.data.exc import DataNotFoundError
-from avengers.data.models.user import UserModel
-from avengers.data.repositories.user import UserRepository
 from avengers.data.models.unauthorized_user import UnauthorizedUserModel
+from avengers.data.models.user import UserModel
+from avengers.data.repositories.email import send_mail
+from avengers.data.repositories.refresh_token import RefreshTokenRepository
 from avengers.data.repositories.unauthorized_user import UnauthorizedUserRepository
+from avengers.data.repositories.user import UserRepository
 from avengers.presentation.exceptions import UserAlreadyExists, InvalidSignupInfo, SignupAlreadyRequested, \
     InvalidVerificationKey, UserNotFound, TokenError
 
@@ -31,33 +31,36 @@ class AuthService:
         except DataNotFoundError:
             pass
         else:
-            raise UserAlreadyExists("")
+            raise UserAlreadyExists()
 
         already_requested = await self.unauthorized_user_repository.find_by_email(email)
         if already_requested:
-            raise SignupAlreadyRequested("")
+            raise SignupAlreadyRequested()
 
         temp_user = UnauthorizedUserModel(email, generate_password_hash(password))
         saved = await self.unauthorized_user_repository.save(temp_user)
 
-        send_mail(email, title=settings.SIGNUP_EMAIL_TITLE, content=settings.SIGNUP_EMAIL_CONTENT.format(saved))
+        send_mail(
+            email, title=settings.SIGNUP_EMAIL_TITLE, content=settings.SIGNUP_EMAIL_CONTENT.format(saved)
+        )
 
         return json("Verification code and link sent to your email.", 204)
 
     async def verify_key(self, verify_key: str):
         temp_user = await self.unauthorized_user_repository.find_by_uuid(verify_key)
         if not temp_user:
-            raise InvalidVerificationKey("")
+            raise InvalidVerificationKey()
 
+        count_of_users = await self.user_repository.get_count()
         user = UserModel(
             email=temp_user.email,
             password=temp_user.password,
-            receipt_code=None,
-            is_paid=None,
-            is_printed_application_arrived=None,
-            is_passed_first_apply=None,
-            is_passed_interview=None,
-            is_final_submit=None,
+            receipt_code=count_of_users,
+            is_paid=False,
+            is_printed_application_arrived=False,
+            is_passed_first_apply=False,
+            is_passed_interview=False,
+            is_final_submit=False,
             exam_code=None,
             volunteer_score=None,
             attendance_score=None,
@@ -77,10 +80,10 @@ class AuthService:
         try:
             saved = await self.user_repository.get(email)
         except DataNotFoundError:
-            raise UserNotFound("")
+            raise UserNotFound()
 
         if not check_password_hash(saved.password, password):
-            raise UserNotFound("")
+            raise UserNotFound()
 
         is_refresh_saved = await self.refresh_token_repository.get(email)
         if is_refresh_saved:
@@ -100,7 +103,7 @@ class AuthService:
 
         email = await self.refresh_token_repository.get_by_refresh(refresh)
         if not email:
-            raise TokenError("")
+            raise TokenError()
 
         access = await create_access_token(identity=email, app=request.app)
         return json({"access": access}, 201)
@@ -113,7 +116,7 @@ class AuthService:
 
         saved_refresh = await self.refresh_token_repository.get_by_refresh(refresh)
         if not saved_refresh:
-            raise TokenError("")
+            raise TokenError()
 
         await self.refresh_token_repository.delete(saved_refresh)
         return json(dict(msg="Logout succeed"), 204)
